pyproject.toml
[project]
name = "livraria_base"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "Marco André Mendes", email = "marcoandre@gmail.com"},
]
dependencies = [
    "django>=5.1.1",
    "djangorestframework>=3.15.2",
]
requires-python = "==3.12.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
